requirements.txt
pandas
numpy
scikit-learn
matplotlib
xgboost
seaborn
streamlit
pyyaml
pillow
altair